sdhci: support mmc host controller(sdhci-of-dwcmhc)
Support to set max_req_size and max_regs in dts.
The status of this node defaults to disable becaues the controller does not exist in most of bitstream.
If you want to use this mmc host,use /nfs/home/guoyaxing/fpga/sd_2_10/ and set status = "okay" in this node

Signed-off-by: Guo Yaxing <guoyaxing@bosc.ac.cn>